Use windows-sys for vmbus_proxy.
